[ZEPPELIN-2291] Notebook commit dropdown cut-off
### What is this PR for?
When you have a notebook under GIT version control, the dropdown with your commits will be cut-off with no possibility to scroll to older commits (see attached screenshot).

This fixes it by applying a relative max-height and appropriate overflow (similar styles as for the notebooks dropdown in the navigation bar). If the dropdown is longer than max-height, a scrollbar will be shown :)

### What type of PR is it?
Bug Fix

### What is the Jira issue?
https://issues.apache.org/jira/browse/ZEPPELIN-2291

### How should this be tested?
Create a notebook under version control and make a couple of changes & commits, until the dropdown gets too long to fit on one screen. If everything works as intended, you should now be able to scroll to older commits inside the dropdown.

### Screenshots (if appropriate)
![apache_zeppelin_commit_dropdown_cut_off](https://cloud.githubusercontent.com/assets/1144202/24155455/3041dcfc-0e54-11e7-8120-3617d579f29f.png)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Alexander Wenzel <wenzel@myhammer.de>

Closes #2173 from elkangaroo/ZEPPELIN-2291-Commit-Dropdown-Scroll and squashes the following commits:

110df35 [Alexander Wenzel] [ZEPPELIN-2291] Fix cut-off revision dropdown