ZEPPELIN-2421. AngularObject miss AngularObjectListener when it is pushed from frontend
### What is this PR for?
This bug happens when you save your note, and restart zeppelin and run it again. The root cause is that the angular object miss  AngularObjectListener which means you can not sync up the changes on angular object to frontend. This PR would restore the listener correctly when we restore the angular objects saved before.  Also make some changes for logging in this PR.

### What type of PR is it?
[Bug Fix]

### Todos
* [ ] - Task

### What is the Jira issue?
* https://issues.apache.org/jira/browse/ZEPPELIN-2421

### How should this be tested?
Tested manually

### Screenshots (if appropriate)

Before
![zeppelin_before](https://cloud.githubusercontent.com/assets/164491/25173268/0824dc72-2526-11e7-955c-ea5e20dae746.gif)

After
![zeppelin_after](https://cloud.githubusercontent.com/assets/164491/25173272/09e11bfc-2526-11e7-9096-5adb6f1992ef.gif)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Jeff Zhang <zjffdu@apache.org>

Closes #2263 from zjffdu/ZEPPELIn-2421 and squashes the following commits:

c2503a4 [Jeff Zhang] ZEPPELIN-2421. AngularObject miss AngularObjectListener when it is pushed from frontend