added principalSuffix to AD examples
### What is this PR for?
Documentation - Adding an example use of prinicpalSuffix to the Active Directory authentication documentation. It took us a while to figure out that the lack of a principalSuffix was our problem and then a bit more hunting to figure out how to automatically include it

### What type of PR is it?
Improvement

### Todos
* [ ] - Task

### What is the Jira issue?
* Open an issue on Jira https://issues.apache.org/jira/browse/ZEPPELIN/
* Put link here, and add [ZEPPELIN-*Jira number*] in PR title, eg. [ZEPPELIN-533]

### How should this be tested?
Outline the steps to test the PR here.

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update?
* Is there breaking changes for older versions?
* Does this needs documentation?

Author: Paul Brenner <paul-brenner@users.noreply.github.com>

Closes #2281 from paul-brenner/patch-1 and squashes the following commits:

2ff57a9 [Paul Brenner] added principalSuffix to AD examples

(cherry picked from commit ec84613657f6494e8a265ee3713a6ef1751115c6)
Signed-off-by: Lee moon soo <moon@apache.org>