[ZEPPELIN-2463] Avoid Locking interpreterSettings during Notebook deletion
Deletion of a notebook requires locking interpreterSettings. If the deletion is delayed , then lock is not released. At that point, we cannot run any notebook because everything is waiting to lock interpreterSettings.
Looking at the code, there is no reason to lock the InterpreterSettings object in InterpreterSettingManager.removeNoteInterpreterSettingBinding.
Similarly in InterpreterSettingManager.getNoteInterpreterSettingBinding only interpreterSettingBinding is accessed and its already a thread safe object. So we can remove synchronization on InterpreterSettings

Bug Fix

https://issues.apache.org/jira/browse/ZEPPELIN-2463

Being a concurrency issue, it is difficult to test.
Please see the comments in the jira to see the issue experienced on a production zeppelin server.

* Does the licenses files need update? NO
* Is there breaking changes for older versions? NO
* Does this needs documentation? NO

Author: Benoy Antony <benoy@apache.org>

Closes #2296 from benoyantony/ZEPPELIN-2463 and squashes the following commits:

6f28d74 [Benoy Antony] replace check and remove with remove
0e0bb60 [Benoy Antony] fix checkstyle warnings
67f1514 [Benoy Antony] ZEPPELIN-2463 Avoid Locking interpreterSettings during Notebook deletion

(cherry picked from commit d37684403af045b343b1008977bcf6deef06e1d5)
Signed-off-by: Lee moon soo <moon@apache.org>