[HOTFIX] Set interpreter context in zeppelin context from SparkSqlInterpreter
### What is this PR for?
When running `%sql` paragraphs sometimes a NullPointer is thrown instead of presenting the result.

This was traced back to `sparkInterpreter.getZeppelinContext().setInterpreterContext(context);` not being called as expected from SparkSqlInterpreter. Since the SparkInterpreter is used from the SparkSqlInterpreter, if a `%spark` paragraph is executed before a `%sql` paragraph it works as expected (as SparkInterpreter sets the interpreter context in the zeppelin context).

### What type of PR is it?
[Hot Fix]

### Todos

### What is the Jira issue?
Hotfix

### How should this be tested?
Start the spark context by executing
```
%sql
select 1
```
Expected result is a table with 1 in.
Before this patch a NullPointer will be thrown.

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: David Ersson <dersson@kth.se>

Closes #2334 from FireArrow/sqlinterpreter_hotfix and squashes the following commits:

3d2b0a9 [David Ersson] Set interpreter context in zeppelin context