[ZEPPELIN-2469] HeliumBundleFactoryTest fails on Travis
### What is this PR for?
There are 6 errors occuring with error: "yarn config set cache-folder /tmp/ZeppelinLTest_1494573109020/helium-bundle/yarn-cache --registry=http://registry.npmjs.org/' failed."
The failures seem to occur because the "yarnCacheDir" directory is not created. For this, I have added code to create the directory before the absolute path is stored in yarnCacheDirPath.

### What type of PR is it?
Bug Fix

### What is the Jira issue?
https://issues.apache.org/jira/browse/ZEPPELIN-2469

### How should this be tested?
The tests should pass on the CI

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: pravin-dsilva <pravind@us.ibm.com>

Closes #2362 from pravin-dsilva/ZEPPELIN-2469 and squashes the following commits:

19fd231 [pravin-dsilva] update travis.yml
1939013 [pravin-dsilva] [ZEPPELIN-2469] HeliumBundleFactoryTest fails on Travis