[ZEPPELIN-2471] - NotebookTest fails on travis
### What is this PR for?
The stacktrace for this issue gives the error:
```
06:31:15,263 ERROR org.apache.zeppelin.notebook.repo.VFSNotebookRepo:151 - Can't read note file:///tmp/ZeppelinLTest_1495261875233/notebook/2BQA35CJZ
com.google.gson.JsonSyntaxException: 2016-03-29T16:21:09-0700
```
This issue is related to notebooks failing to load with date ParseException and  was fixed in ZEPPELIN-1129. However, in ZEPPELIN-2395 the fix for old format is excluded and hence the issue occurs again. Added the same code to fix the old date format issue in the fromJSON method of Note.java

### What type of PR is it?
Bug Fix

### What is the Jira issue?
https://issues.apache.org/jira/browse/ZEPPELIN-2471

### How should this be tested?
Tests should pass on CI

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: pravin-dsilva <pravind@us.ibm.com>

Closes #2366 from pravin-dsilva/ZEPPELIN-2471 and squashes the following commits:

3be65aa [pravin-dsilva] Update indentation for Note.java
d083208 [pravin-dsilva] Update .travis.yml to include NotebookTest
3a2e788 [pravin-dsilva] ZEPPELIN-2471 - NotebookTest fails on travis