Improved comment, removed admin user definition per default
### What is this PR for?
Improved shiro.ini.template.

Since the file is just a template, it should be such that if the user copies it from shiro.ini.template to shiro.ini the auth settings are comparable conservative.
Hence: The admin user has been commented out to avoid enabling admin by
default with a default password.

In addition, the comment in the url section was misleading. It has been clarified. The comment "To enfore security, comment the line below and uncomment the next one“ was
misleading, since the setting was already on "authc" and it was not the "line below". This has been fixed.

### What type of PR is it?
[ Improvement]

### Todos
* N/A

### What is the Jira issue?
* https://issues.apache.org/jira/browse/ZEPPELIN-2588

### How should this be tested?
Non regression test / unit testing.

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? *NO*
* Is there breaking changes for older versions? *NO*
* Does this needs documentation? *NO*

Author: Christian Fries <email@christian-fries.de>

Closes #2368 from cfries/improved-shiro-ini and squashes the following commits:

e7606497 [Christian Fries] Improved comment, removed admin user definition per default