0.7.1 with ssl
### What is this PR for?
The Cassandra Interpreter does not support talking to clusters that use SSL/client to node encryption. It does not have the properties needed to configure the SSL Context. This PR adds the properties to the driver config and sets up the SSL options when they are requested.

### What type of PR is it?
Feature

### What is the Jira issue?
[ZEPPELIN-1501](https://issues.apache.org/jira/browse/ZEPPELIN-1501)

### How should this be tested?
Using Cassandra/CQL interpreter connect to a Cassandra cluster that uses a client to node encryption, i.e:

```
%cassandra
describe keyspaces;
```
### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? Yes.
    The new SSL-related properties should be added to the list (cassandra.ssl.enabled, cassandra.ssl.truststore.path and cassandra.ssl.truststore.password).

Author: Alex Lourie <djay.il@gmail.com>
Author: Robert Marshall <robertmarshall@Instaclustrs-MacBook-Pro.local>

Closes #2238 from alourie/0.7.1-withSSL and squashes the following commits:

2a6eee6 [Alex Lourie] Handle potential NPE in properties parsing
e9b4769 [Alex Lourie] Fix formatting for CI to pass
521b2d9 [Robert Marshall] Add SSL support to Cassandra interpreter