[ZEPPELIN-2885] Have Logger subclass StringIO and override write method
H2O python module uses the `isatty` method when outputting to stdout. As Zeppelin replaces stdout with a custom Logger class, the lack of this method causes H2O to fail with most operations.

Subclassing StringIO and overriding the write method fixes this ensures other behavior in the TextIO classes used for stdout is present as well, obviating the need to keep adding method stubs to workaround other modules in the future.

[Bug Fix]

* [ ] - Does `reset` still need to be present?

* https://issues.apache.org/jira/browse/ZEPPELIN-2885

* Python/Pyspark interpreter output still works
* Running the following paragraph should return `False`

```
%python
import sys
sys.stdout.isatty()
```

* Does the licenses files need update?			No
* Is there breaking changes for older versions?		No
* Does this needs documentation?      			No