[ZEPPELIN-2880] - Fix username output when OIDC is enabled
### What is this PR for?
When OIDC is enabled, user's roles/permissions/tokens are sent to Zeppelin's client via websocket and appears in the web browser when running a paragraph in a notebook.

### What type of PR is it?
[Bug Fix]

### What is the Jira issue?
* [ZEPPELIN-2880](https://issues.apache.org/jira/browse/ZEPPELIN-2880)

### How should this be tested?
Enable OIDC , login to Zeppelin and run a paragraph.  Check the text "Last updated by".

### Questions:
* Does the licenses files need update?  N/A
* Is there breaking changes for older versions?  N/A
* Does this needs documentation?  N/A

Author: byamthev <byamthev@gmail.com>

Closes #2551 from byamthev/zeppelin_2880 and squashes the following commits:

cb6ec70 [byamthev] [ZEPPELIN-2880] - Fix username output when OIDC is enabled
11d0729 [byamthev] [ZEPPELIN-2880] - Fix username output when OIDC is enabled
7263689 [byamthev] [ZEPPELIN-2880] - Fix username output when OIDC is enabled