[ZEPPELIN-3060] Unable to use interpreter names with special characters (e.g. dash or underscore)
### What is this PR for?
This fixes a regression which was introduced with ZEPPELIN-3013. There
is no restriction regarding characters when creating new interpreters.
But when running a note the interpreter name was parsed using the regex
%([\w\.]+).*, which only considered the word character class and dots.
Interpreter names with special characters (e.g. _ or -) were not matched
correctly. This change enables usage of interpreter names that use other
characters by changing the regex group to %([^\s\(]+).*, which means:
parse the interpreter name from % until any whitespace or open
parenthesis.

### What type of PR is it?
Bug Fix

### Todos
-

### What is the Jira issue?
https://issues.apache.org/jira/browse/ZEPPELIN-3060

### How should this be tested?
Create an interpreter with a name containing one or more special
characters, e.g. dash or underscore. Try running that interpreter.

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? no 
* Is there breaking changes for older versions? no
* Does this needs documentation? no