[ZEPPELIN-3076]Chart field is also draggable and sortable in the 'keys', 'groups' and 'values'
### What is this PR for?
Current the `keys`, `groups` and `values` is only droppable. It is a little inconvenient if I want to drag it to other place or sort it.
This feature let `keys`, `groups` and `values` not only `droppable`, but also `draggable` and `sorttable`.

### What type of PR is it?
[Feature]

### What is the Jira issue?
https://issues.apache.org/jira/browse/ZEPPELIN-3076 [ZEPPELIN-3076]

### How should this be tested?
I have Add the Screenshots for the test. Just as the Screenshots showing, all the element in `keys`, `groups` and `values` is draggable and sorttable.

![untitled project4](https://user-images.githubusercontent.com/5969176/33252922-ed1ab0b4-d37b-11e7-8a5c-b3dbb6765d18.gif)

### Questions:
* Does the licenses files need update? NO
* Is there breaking changes for older versions? NO
* Does this needs documentation? NO

Author: liguohui <liguohui@cmss.chinamobile.com>

Closes #2686 from liguohuicmss/draggable-field-keys-groups-values2 and squashes the following commits:

08b86b8 [liguohui] add a new line
ebb8436 [liguohui] Chart field is also draggable and sortable in the 'keys', 'groups' and 'values'