[ZEPPELIN-3098] Livy Interpreter fails if row contains newline or tab
### What is this PR for?
This PR fixes the issue of newlines and tabs breaking results in the SQL interpreter in Livy.

The Livy interpreter will return incorrect results if a row contains \n or \t characters.
In the case of the newline, the result will be:
Line is missing from results if the \n appears anywhere but the end of a cell
`String index out of range: 17` if it appears at the end of a cell
In the case of the tab, the result will be misaligned columns if the tab appears in the middle of a cell
The output showing these error is attached to the Jira.

I have changed the parsing and any newline or tab characters will be escaped

### What type of PR is it?
Bug Fix

### What is the Jira issue?
https://issues.apache.org/jira/browse/ZEPPELIN-3098

### How should this be tested?
Unittests have been added

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Alex Bush <alex.bush@coxauto.co.uk>
Author: Alex Bush <bushnoh@users.noreply.github.com>

Closes #2701 from bushnoh/zeppelin-3098 and squashes the following commits:

9f2f6e3 [Alex Bush] Merge pull request #3 from apache/master
14e120d [Alex Bush] Merge pull request #2 from apache/master
46981fd [Alex Bush] Merge pull request #1 from apache/master
ee5a41b [Alex Bush] Merge remote-tracking branch 'upstream/master'
67a93b5 [Alex Bush] Merge branch 'master' of https://github.com/bushnoh/zeppelin
31cdbdc [Alex Bush] Added another comment explaining the regexp change
d054af0 [Alex Bush] Force a dummy change for Travis
64a42be [Alex Bush] Fix for newline and tab in data