[ZEPPELIN-3263] Prevent login window from popping up on logout
When our ticket is no longer valid, server sends a session_logout
message, so that front-end could prompt user to re-login. This is
handy when ticket expires by itself, but is not needed when user
explicitly logs out.

Whether or not login window will appear is not deterministic - it
depends on whether or not there were requests in flight during logout.
This creates problems for AuthenticationIT testcase, and may confuse
the users.

This change prevents login window from appearing after explicit logout.