ZEPPELIN-3391. Incorrect status shown for '%livy2.conf' and %spark2.conf' interpreters
### What is this PR for?
The root cause of this bug is that the final status of paragraph is not only determined by interpreter, but also by its scheduler. Here the bug is in FIFOScheduler. I just fix it as a quick solution in this PR. But for a long term solution, I think we may need to do code refactoring in the scheduler component.

### What type of PR is it?
[Bug Fix]

### Todos
* [ ] - Task

### What is the Jira issue?
* https://issues.apache.org/jira/browse/ZEPPELIN-3391

### How should this be tested?
* Manually tested

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Jeff Zhang <zjffdu@apache.org>

Closes #2926 from zjffdu/ZEPPELIN-3391 and squashes the following commits:

09441ad [Jeff Zhang] ZEPPELIN-3391. Incorrect status shown for '%livy2.conf' and %spark2.conf' interpreters

(cherry picked from commit 0e82a5712ea6229925b7b9663d492b54d3c29b59)
Signed-off-by: Jeff Zhang <zjffdu@apache.org>