[ZEPPELIN-3522] String "defaultValue" instead of boolean in some "interpreter-settings.json" files
### What is this PR for?
The _interpreter-settings.json_ file for each interpreter has details of each configurable parameter for that interpreter. Each parameter also has a _defaultValue_ setting. For boolean-typed parameters the _defaultValue_ must be set to _true_ or _false_.

But in some _interpreter-settings.json_ files, the _defaultValue_ has been set to the string values _"true"_ or _"false"_ (the quote marks are included in the value provided).

### What type of PR is it?
[Bug Fix]

### Todos
* [ ] - Task

### What is the Jira issue?
https://issues.apache.org/jira/browse/ZEPPELIN-3522

### How should this be tested?
CI pass

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Sanjay Dasgupta <sanjay.dasgupta@gmail.com>

Closes #2998 from sanjaydasgupta/zepp-3522-string-default-value and squashes the following commits:

9bdb0f35b [Sanjay Dasgupta] zepp-3522-string-default-value: Initial updates