[ZEPPELIN-3528] Ordering by interpreter name in same interpreter group
### What is this PR for?
In 'Create new note' diagram, 'Default Interpreter' is ordered by only interpreter groups.

But interpreter is not ordered in same interpreter group.
(It seems to be ordered by hashcode of 'InterpreterSetting')

When there are many interpreter in same group and this group is first,
user cannot predict what interpreter is showed first.

So I think that interpreter need to be ordered by name.

### What type of PR is it?
[Improvement]

### Todos
* [x] - Modify Code

### What is the Jira issue?
* https://issues.apache.org/jira/browse/ZEPPELIN-3528

### How should this be tested?
* Click 'create new note'

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update?
    * No
* Is there breaking changes for older versions?
    * No
* Does this needs documentation?
    * No

Author: Jaehoon Jeong <jaehoon.jeong@woowahan.com>

Closes #3005 from woowahan-jaehoon/ZEPPELIN-3528 and squashes the following commits:

5cb709ce3 [Jaehoon Jeong] [ZEPPELIN-3528] Ordering by interpreter name in same interpreter group